feat: implement validation utilities in validateRequest.ts
