--- server/utils/validateFormData.ts
@@ -1,19 +1,0 @@
-import type { z } from 'zod'
-
-export default async <T extends z.ZodTypeAny>(
-    schema: T,
-    transformer?: (formData: FormData) => Record<string, unknown>
-): Promise<z.infer<T>> => {
-    const formData = await readFormData(useEvent())
-
-    // デフォルトはFormDataをオブジェクトに変換
-    const dataToValidate = transformer
-        ? transformer(formData)
-        : Object.fromEntries(formData.entries())
-
-    const result = schema.safeParse(dataToValidate)
-
-    if (!result.success) throw result.error
-
-    return result.data
-}
