--- server/utils/validateParams.ts
@@ -1,13 +1,0 @@
-import type { z } from 'zod'
-
-export default async <T extends z.ZodTypeAny>(
-    schema: T
-): Promise<z.infer<T>> => {
-    const result = await getValidatedRouterParams(useEvent(), (body) =>
-        schema.safeParse(body)
-    )
-
-    if (!result.success) throw result.error
-
-    return result.data
-}
